+++ src/pages/Study/OldStudy.tsx
@@ -1,5 +1,5 @@
 import React, { Fragment } from 'react';
-import StudyItem from './StudyItem';
+import StudyAccordion from './share/StudyAccordion';
 
 interface OldStudyProps {
   studyList: number[];
@@ -11,7 +11,7 @@ const OldStudy = ({ studyList }: OldStudyProps) => {
       {/* TODO 추후 게시판이 구현되면 컴포넌트를 그대로 가져와서 사용할 예정!! */}
       {studyList?.map((study) => (
         <Fragment key={study}>
-          <StudyItem id={study} />
+          <StudyAccordion />
         </Fragment>
       ))}
     </div>
